--- stweb/src/components/Tournament/constants.js
@@ -1,3 +1,0 @@
-export const SHOW_RECORD = "show";
-export const EDIT_RECORD = "edit";
-export const REMOVE_RECORD = "remove";
